Moved click stuff to __main__.py
