+++ practical_porcupines/__main__.py
@@ -1,4 +1,57 @@
-from click_cli import base_group
+import click
+from utils import ConfigBase, ConfigApi, ConfigWebPortal, ConfigBot
+from discord_bot import bot_client
+from flask_api import flask_api_app
+from flask_webportal import flask_webportal_app
+
+
+@click.group()
+def base_group():
+    pass
+
+
+@click.command()
+def flask_webportal():
+    """
+    Runs the webportal mini-project (like the discord bot but on a website ui)
+    """
+
+    config_webportal = ConfigWebPortal()
+
+    flask_webportal_app.run(
+        host=config_webportal.API_DOMAIN,
+        port=config_webportal.API_PORT,
+        debug=ConfigBase().SHOULD_DEBUG
+    )
+
+
+@click.command()
+def discord_bot():
+    """
+    Runs the discord bot mini-project, displaying all from the flask_api mini-project
+    """
+
+    pass
+
+
+@click.command()
+def flask_api():
+    """
+    The core API of this project, RESTFUL and built with flask & flask-restful
+    """
+
+    config_api = ConfigApi()
+
+    flask_api_app.run(
+        host=config_api.API_DOMAIN,
+        port=config_api.API_PORT,
+        debug=ConfigBase().SHOULD_DEBUG
+    )
+
+
+base_group.add_command(flask_api)
+base_group.add_command(discord_bot)
+base_group.add_command(flask_webportal)
 
 if __name__ == "__main__":
     base_group()
